Debug such that blackjack.py runs smoothly
First player was getting blackjack by default due to the if-statement in
sum_cards() being passed a fact rather than a conditional. This has been
corrected, and now the correct score is calculated, and the winning player
or players are acknowledged.

Addressed incomplete copy-paste from BlackjackGame.deal() to
BlackjackGame.hit().

Changed tabs to three spaces in printed messages, for a cleaner look.