Make "sum_hand()" more pythonic...
Iterating through a list via its indices is not pythonic, but iterating
through its values does not update the list in its place. As a solution,
I create an empty list "intvals," pop() each item off of vals, edit it if
needed, and append it to intvals. The total number of items between vals
and intvals is constant at the end of each iteration, and so no significant
undue memory load is imposed by introducing a second list.