Extended discard functionality
Extended the __init__ method and adjusted some other methods to allow the user to specify another Deck object to act as the discard pile.

Helpful if you want to be able to manipulate the discard pile without accessing the _discard_pile, whilst maintaining the functionality shuffle_back(), discarded, etc.

Also added a unit test for this new functionality.